update deepspeed ckpt load

+++ configs/opensora/inference/16x256x256.py
@@ -28,7 +28,7 @@ scheduler = dict(
     cfg_scale=7.0,
     cfg_channel=3,  # or None
 )
-dtype = "fp16"
+dtype = "bf16"
 
 # Others
 batch_size = 1
